requirements.txt
sentence-transformers>=2.2.2
langchain>=0.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0 